requirements.txt
pyshark
pandas
openpyxl
mysql-connector-python
numba
line_profiler
pyahocorasick
regex
aiofiles